Updated MDI to be loaded locally
I had to kind of work around the fact that the github repo
for material-design-icons is outdated; some icons do not show up.
I found a recent Issue and downloaded the most recent version from
Google's hosted files for the material icon font. Even so, some icons
don't work, like battery level icons.

Here's the link to that Issue on the MDI repo:
https://github.com/google/material-design-icons/issues/786