Added Graph toggle
I started implementing a toggle for the graph on the Overview page. I
want the user to be able to toggle between different graphs using
buttons or a selection dropdown as a toggle. Right now I'm using buttons
and it more or less works. I still need to figure out some stuff on the
Overview component to actually send it different data as props. That
shouldn't be too hard, just need to use the selectedGraph state which
is wired up to get updated properly.