Fixed #7
When toggling between graphs, the last data node is no longer
duplicated. Since the props are always being sent, I needed to prevent
them from updating the state.

However, I noticed new behavior that I'm not sure we want: when you
toggle between the graphs, of course, the data for that graph isn't
being updated unless it's in view. This is probably more performance
friendly, but we might want it to be cached/updating in the background
while it's not in view. One idea for changing this would be to change
how toggling works from conditionally mounting the correct Chart
component, to simply render both of them, but hide/unhide them using CSS
to only have one in view, but both should theoretically be updating
while they're invisible.