Fixed #9 and reorganized graph wrapper(s)
Now when you toggle between graphs, the data is updated even when its
not in view. This means both/all graphs are always being updated, but
only one is visible at a time.

I also reorganized the nesting of the divs containing the actual charts.

In doing this, I realized that a more robust "chart framework" would
be beneficial, but I think I'll hold off on that until I have more time.
This works, but making the graphs more modular would get us less code
duplication, and a bit easier management. This would be even more
useful on the Analytics view, which will include multiple graphs/more
graphing options. I think it might be worth working on sooner than later
just for that view.