Started implementing the graph framework
First of all, I created a new file called `graph-config.js`, which
simply holds the config for all graphs associated with the
specific parent. Right now I'm importing it into the Graph component,
but I would rather move that up to the Parent page/view component, and
send it down as a prop.

I also temporarily commented out the functionality for updating the
graphs, because this needs to be reworked for the new references to the
graph config. I also want to try to make that and a helper function work
for any number of graphs that the `graph-config.js` file has in it.