Made progress on #21
I added a ternary operator that chooses to generate the correct chart
type based on the graphs `type` attribute which is now to be set for a
graph the `graph-config.js` at the same level as `name`.

I'm not calling this one fixed yet, because I want to move the Graph
component to a universal location in the file tree, and also add all the
graph types. Right now I just have 'line' and 'bar' types working.

The other graph types should be fine, it's just a matter of the naming
and knowing what `react-chartjs-2` component it is. For example and area
chart is actually just a line or radar graph with a different dataset
option.