Fixed #22 – net gauge (horizontal bar chart) is in
This proved that the "graph system" worked well for the graphs as they
were, but was not as extensible as I had hoped in its state.

For example, I had to declare a `set1` object and initialize it with
the graphs I wanted to display just for the first graph. Then another
variable, `set2` with just the second graph I wanted to display below
the first Graph component. This wasn't hard, but it was an oversight
(it's something I had thought about previously, and thought I took into
account, but I was thinking the `graphSet` is an array, which it is
not but rather an object).

There were other things I had to fix to make the Net Power "gauge" to
work as intended, since its a sort of specialty chart (I'm using a
horizontal bar chart in a way that resembles a horizontal gauge).

Long story short, `Graph.js` as a universal component can't happen
and the whole "graph system" is becoming less useful. Nice!