Worked on Array page (#17)
I started setting up the widgets for the Array page with the proper
values. The graphs need to be setup for the specific array set of graphs
which includes tweaking the Graph component and changing the
graph-config file.

I also fixed a bug where the title of a graph didn't update after the
first changed selection (this code was duplicated across all the other
parent view components, so it was fixed in all those as well).